requirements/base.txt
Django==1.8.4
Unipath=1.1
django-model-utils=2.3.1